% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_hclust.R
\name{as_hclust}
\alias{as_hclust}
\title{Coerce Input To \code{hclust}}
\usage{
as_hclust(
  x,
  distfun = stats::dist,
  linkagefun = function(d) stats::hclust(d, method = "complete")
)
}
\arguments{
\item{x}{A numeric matrix/data frame, \code{dist}, or \code{hclust} object.}

\item{distfun}{Distance function used when \code{x} is matrix-like.}

\item{linkagefun}{Linkage function used when \code{x} is matrix-like or \code{dist}.}
}
\value{
An \code{hclust} object.
}
\description{
Coerce Input To \code{hclust}
}
