% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_branches.R
\name{default_colorscale}
\alias{default_colorscale}
\title{Default Dendrogram Colorscale}
\usage{
default_colorscale()
}
\value{
Character vector of 8 RGB colors.
}
\description{
Default Dendrogram Colorscale
}
