% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendro_layout.R
\name{dendro_layout}
\alias{dendro_layout}
\title{Build Plotly Layout For Dendrogram Data}
\usage{
dendro_layout(dendro, orientation = "bottom")
}
\arguments{
\item{dendro}{A list with \code{segments}, \code{labels}, and \code{nodes}.}

\item{orientation}{One of \code{"bottom"}, \code{"top"}, \code{"left"}, or \code{"right"}.}
}
\value{
A plotly layout list.
}
\description{
Build Plotly Layout For Dendrogram Data
}
