% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_dependency.R
\name{.locate_dendro_asset_dir}
\alias{.locate_dendro_asset_dir}
\title{Locate The plotly-dendro JS Assets}
\usage{
.locate_dendro_asset_dir()
}
\value{
Absolute path to the asset directory.
}
\description{
Supports installed-package usage and development workflows loaded via
\code{pkgload::load_all()}, where \code{system.file(package = ...)} can be empty.
}
\keyword{internal}
