% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orient_data.R
\name{orient_data}
\alias{orient_data}
\title{Orient Dendrogram Coordinates}
\usage{
orient_data(dendro, orientation = "bottom")
}
\arguments{
\item{dendro}{A list with \code{segments}, \code{labels}, and \code{nodes}.}

\item{orientation}{One of \code{"bottom"}, \code{"top"}, \code{"left"}, or \code{"right"}.}
}
\value{
The transformed \code{dendro} list.
}
\description{
Orient Dendrogram Coordinates
}
