% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dendro_nodes.R
\name{add_dendro_nodes}
\alias{add_dendro_nodes}
\title{Add Dendrogram Merge Node Layer}
\usage{
add_dendro_nodes(
  p,
  dendro,
  orientation = "bottom",
  ...,
  data = NULL,
  inherit = FALSE
)
}
\arguments{
\item{p}{A plotly object.}

\item{dendro}{A \code{dendro_data()} result.}

\item{orientation}{Dendrogram orientation.}

\item{...}{Passed to \code{plotly::add_trace()}.}

\item{data}{Optional plotly data.}

\item{inherit}{Plotly inheritance flag.}
}
\value{
A plotly object.
}
\description{
Add Dendrogram Merge Node Layer
}
