% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dendro_segments.R
\name{add_dendro_segments}
\alias{add_dendro_segments}
\title{Add Dendrogram Segments Layer}
\usage{
add_dendro_segments(
  p,
  dendro,
  orientation = "bottom",
  color_threshold = NULL,
  colorscale = NULL,
  ...,
  data = NULL,
  inherit = TRUE
)
}
\arguments{
\item{p}{A plotly object.}

\item{dendro}{A \code{dendro_data()} result.}

\item{orientation}{Dendrogram orientation.}

\item{color_threshold}{Height cutoff for branch colors.}

\item{colorscale}{Optional color vector.}

\item{...}{Passed to \code{plotly::add_segments()}.}

\item{data}{Optional plotly data.}

\item{inherit}{Plotly inheritance flag.}
}
\value{
A plotly object.
}
\description{
Add Dendrogram Segments Layer
}
