% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_branches.R
\name{color_branches}
\alias{color_branches}
\title{Color Dendrogram Segments}
\usage{
color_branches(segs, hc, color_threshold = NULL, colorscale = NULL)
}
\arguments{
\item{segs}{Segment data frame.}

\item{hc}{An \code{hclust} object.}

\item{color_threshold}{Height cutoff for color grouping.}

\item{colorscale}{Optional color vector.}
}
\value{
\code{segs} with a \code{color} column.
}
\description{
Color Dendrogram Segments
}
