% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_labels.R
\name{extract_labels}
\alias{extract_labels}
\title{Extract Leaf Labels}
\usage{
extract_labels(hc, labels = NULL)
}
\arguments{
\item{hc}{An \code{hclust} object.}

\item{labels}{Optional custom labels.}
}
\value{
A data frame with columns \code{x}, \code{y}, \code{label}.
}
\description{
Extract Leaf Labels
}
